[feature] UsageScreen: finalise layout for compact screen width
